moved pipette movement conflict to separate file
